+++ packages/xrpl/HISTORY.md
@@ -6,6 +6,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ### Added
 * Support for `NFTokenMintOffer` (XLS-52)
+* Add support for `Batch` amendment (XLS-56)
 
 ### Fixed
 * `OracleSet` transaction accepts hexadecimal string values for `AssetPrice` field
@@ -20,7 +21,6 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 * Support for the Permissioned Domains amendment (XLS-80).
 * Support for the `simulate` RPC ([XLS-69](https://github.com/XRPLF/XRPL-Standards/tree/master/XLS-0069-simulate))
 * Support for XLS-77d Deep-Freeze amendment
-* Add support for `Batch` amendment (XLS-56)
 
 ### Changed
 * Deprecated `setTransactionFlagsToNumber`. Start using convertTxFlagsToNumber instead